refactoring - move data to BerlinBucketListItem
Both Home screen and Recommendations screen contains items (cards) that have exactly the same structure, so instead of having Category and BerlinBucketListItem as separate data classes, it makes sense to have just the BerlinBucketListItem with nullable properties - it allows us to avoid code duplication.

This way we can use BerlinBucketListItem for the elements in all the app screens - the structure of the elements is identical in almost all of them, so this is acceptable for this application.