update place uiState once we choose the category
This is needed, because when we use Expanded screen, we show both categories and details lists - in this case the details list state should already be updated in order to be shown (otherwise the app crashes).

Thus, once a category is chosen, we update the details state with the data of the very first place in the list of all places that belong to the chosen category.